pyproject.toml
[tool]
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "liquidsnake", email = "bentsoft365@gmail.com"},
]
dependencies = [
    "django>=4.2.2",
    "djangorestframework>=3.14.0",
    "python-decouple>=3.8",
    "csv-wizard>=1.0.0.post1",
]
requires-python = ">=3.11"
license = {text = "MIT"}
